==> 2_clean_data_sf_crime.ipynb <==
# Clean San Francisco Crime data 
- Jim Haskin

- GA-Data Science
- Dec 2015

- 2/17/2016

## Method
- I have collected the incident reports of the San Franciso Police Department from the SF OpenData website. https://data.sfgov.org/data?category=Public%20Safety. I have the records from January, 2003 until the beginning of 2016.
- I cleaned and reformated the fields.
- I summerized the report to generate a daily report of the number of incidents and another factor I am calling Crime Level. Each incident is given a score based on how violent it is. Murders and assaults are rated high. Traffic violations and non-criminal incidents are rated low. These scores are summed and then normalized to a scale of 0 - 10.
- Also added other crime measuring variables that can be used to narrow the reseach to different subsets of crime.
- 'gun_level', 'COP_count', 'violent_count' are examples. These are explained below.

# Sections


- [Data Source](#Data-Source)
- [Clean Data](#Clean-Data)
- [New Features for Raw Data](#New-Features-for-Raw-data)
- [Corrections for Raw data](#Corrections-for-Raw-data)
- [Create Crime Measuring Variables](#Create-Crime-Measuring-Variables)
- [Consolidate into daily records](#Consolidate-into-daily-records)
- [Normalize Levels](#Normalize-Levels)
- [Write final data to file](#Write-final-data-to-file)
- [qq](#qq)

## Data Source
[[back to top](#Sections)]

- Data downloaded from SF Open Data site. File includes incidents from 1/1/2003 until the present 
- SFPD_Incidents_-_from_1_January_2003.csv
- https://data.sfgov.org/data?category=Public%20Safety


FieldName|Type|Description                             
---------------|------------|---------------------
IncidntNum|string|Police assigned number
Category|string|General Crime category
Descript|string|Secondary category/details
DayOfWeek|string|Day of week event occured
Date|string|Date in format : 01/18/2016
Time|string|Time in format : 23:52
PdDistrict|string|Police District that event occured in
Resolution|int|How case was resolved
Address|string|Address of event
X|float|Longitude 
Y|float|Latitude
Location|string|Latitude,Longitude in character pair
PdId|int|Police Department ID number


import pandas as pd
import numpy as np
import seaborn as sb
%matplotlib inline

! head -2 SFPD_Incidents_-_from_1_January_2003.csv


! tail -2 SFPD_Incidents_-_from_1_January_2003.csv



### Read in Crime data

sf_data = pd.read_csv('SFPD_Incidents_-_from_1_January_2003.csv', index_col=0)    # has header, commas, index
sf_data.head(2)

## Clean Data
[[back to top](#Sections)]

### Convert to lower case
- Feature names
- Feature values that I'm working with

sf_data.columns = sf_data.columns.str.lower()
sf_data['category'] = sf_data['category'].str.lower()
sf_data['descript'] = sf_data['descript'].str.lower()
sf_data['dayofweek'] = sf_data['dayofweek'].str.lower()
sf_data.head(2)

### Investigate data

sf_data.info()

### Observations
- 1,866,570 records 
- date and time in string format
- other fields look appropriate

### Convert date to Pandas datetime format

sf_data['date'] = pd.to_datetime(sf_data['date'])

## New Features for Raw data
[[back to top](#Sections)]

There is information in some of these features that I would like to use in future analysis. But I need to put them into a usable form
- hour - What time did the incident occure. Can be used if you want to split data into time frames.
- month, day, year - Seperate out from data to be used to look for trends.
- shift - feature derived from time that splits the day into 3 working shifts.

### Add the hour as numeric

sf_data['hour'] = sf_data['time'].str[0:2].astype(int)

### Add month, day and year features

#tdf['Date'].dtype
sf_data['month'] = sf_data['date'].dt.month
sf_data['day'] = sf_data['date'].dt.day
sf_data['year'] = sf_data['date'].dt.year
sf_data.head(2)

### Create shift feature
- For more detailed analysis or workforce planning, add feature that records the shift that event occured.
- 3rd shift - Midnight to 7:59am
- 1st shift - 8:00am - 3:59pm
- 2nd shift - 4:00pm - 11:59pm

NOTE: The below discussion of the approximate times in the incident reporting may make this unreliable

def calc_shift(hour):
    shift = hour//8
    if shift == 0:
        shift = 3
    return 'shift_' + str(shift)
        

sf_data['shift'] = sf_data['hour'].apply(calc_shift)
# or leave shift as hour//8. so that it sorts into time order, but label shift0 as third shift
#sf_data['shift'] = sf_data['hour'].apply(lambda x : x//8)

sf_data.head(2)

copy = sf_data.copy()

## Corrections for Raw data
[[back to top](#Sections)]

After analysing the data in the 5_analysis notebook several anomalys appeared. They caused me to remove some data as explained below.

### Investigate Day 1 anomaly
- While investigating the data I found a large spike in the number of incidents on the first day of the month.
- Investigate and reslove

sf_data['day'].value_counts().head()

sf_data[sf_data['day']==15]['time'].value_counts().head()

sf_data['time'].value_counts().head()

sf_data[sf_data['time']=='00:01']['day'].value_counts().head()

sf_data[(sf_data['day']==1) & (sf_data['time']=='00:01')]['time'].count()

### observations
- 6934 records recorded on the the first of the month at 00:01.
- This is ~10% of the records for the 1st and 0.37% of total records.
- My hypotosis is that some incidents were reported later and when the exact date and time were not know, they were recorded this way.
- Since my model is based on the day of the incident, these records are not reliable. 
- If I remove them the counts are more in line with the other days of the month.
- Also appears that many of the times in the records may be approximate. (Many at 12:00). This will not effect the daily totals, but could be a problem if I try to divide the day into shifts.

### observations 2
- there are a total of 47,451 records from all days that have a time of 00:01 ~ 2.5% of data
- These records may be inaccurate also. Since I have 1.8M records it is better to get rid of questionable data

### results
- Remove the records that have a time of 00:01

# Remove day 1 time 00:01
#sf_data = sf_data.drop(sf_data[(sf_data['day']==1) & (sf_data['time']=='00:01')].index)
#sf_data[(sf_data['day']==1) & (sf_data['time']=='00:01')]['time'].count()

# Remove all time 00:01 records
sf_data = sf_data.drop(sf_data[(sf_data['time']=='00:01')].index)
sf_data[(sf_data['time']=='00:01')]['time'].count()

## Create Crime Measuring Variables
[[back to top](#Sections)]

Instead of only calculating the number of incidents per day I have created several other measurements that could be used for analysis and modeling. These can be used to help with workforce planning. These are very subjective and could use domain knowledge to help tune these measurements
- Create crime level feature that weights the incident by the severity/violence of the crime
- Create a weather crime level. I select categories that I feel may be weather related.
- Create a Crime Of Passion feature based on a list of violent/emotional words.
- Create a violent crime feature for the assault, rape and domestic violence categories
- Create a gun related feature


### Investigate the categories and their sub descriptions

sf_data['category'].value_counts()

sf_data[sf_data['category']=='larceny/theft']['descript'].value_counts()

sf_data[sf_data['category']=='other offenses']['descript'].value_counts()

sf_data[sf_data['category']=='non-criminal']['descript'].value_counts()

sf_data[sf_data['category']=='assault']['descript'].value_counts()

sf_data[sf_data['category']=='drug/narcotic']['descript'].value_counts()

sf_data[sf_data['category']=='vandalism']['descript'].value_counts()

sf_data[sf_data['category']=='robbery']['descript'].value_counts()

sf_data[sf_data['category']=='secondary codes']['descript'].value_counts()

sf_data[sf_data['category']=='weapon laws']['descript'].value_counts()

sf_data[sf_data['category']=='prostitution']['descript'].value_counts()

sf_data[sf_data['category']=='sex offenses, forcible']['descript'].value_counts()

sf_data[sf_data['category']=='drunkenness']['descript'].value_counts()

sf_data[sf_data['category']=='disorderly conduct']['descript'].value_counts()

sf_data[sf_data['category']=='family offenses']['descript'].value_counts()

### Assign a Crime level and if Weather related to each catagory
- First number is the crime level in the range 1-4 : Higher number is more violent
- Second number (0/1) indicates if I believe the crime would be effected by the weather

levels = {'larceny/theft' : [2,0],                
          'other offenses' : [1,0],                 
          'non-criminal' : [1,0],
          'assault' :  [4,1],                        
          'vehicle theft' : [2,0],                 
          'drug/narcotic' : [2,0],                 
          'vandalism' : [2,1],                       
          'warrants' : [1,0],                        
          'burglary' : [2,0],                        
          'suspicious occ' : [2,0],                 
          'missing person' : [1,0],                 
          'robbery' : [2,1],                         
          'fraud' : [2,0],                          
          'forgery/counterfeiting' : [2,0],         
          'secondary codes' :  [4,1],              
          'weapon laws' :  [3,1],                    
          'trespass' :  [2,0],                       
          'prostitution' :  [2,0],                  
          'stolen property' :  [2,0],                 
          'sex offenses, forcible' : [4,1],          
          'drunkenness' :  [2,1],                     
          'disorderly conduct'  : [3,1],              
          'recovered vehicle' :  [1,0],              
          'driving under the influence' :  [1,0],      
          'kidnapping' :  [3,0],                      
          'runaway' :  [1,0],                          
          'liquor laws' : [1,0],                     
          'arson' : [3,1],                           
          'embezzlement' : [1,0],                    
          'loitering' : [1,0],                      
          'suicide' :  [1,1],                         
          'family offenses' : [3,1],                  
          'bad checks' : [1,0],                 
          'bribery' : [1,0],                          
          'extortion' : [2,0],                        
          'sex offenses, non forcible' : [2,1],       
          'gambling' : [1,0],                          
          'pornography/obscene mat' : [2,0],          
          'trea' : [1,0]}

### Create new crime variables 'crime_level' and 'weather_crime' from the above mapping

sf_data['crime_level'] = sf_data['category'].map(lambda x : levels[x][0])
sf_data['weather_crime'] = sf_data['category'].map(lambda x : levels[x][1])



### Create new crime variable 'cop' (Crime Of Passion) based on incidents containing the following list of words
- Words were selected from reviewing the incident descriptions looking for words that are more violent or could be triggered by emotions

cop_words = ['assault', 'battery', 'drunk', 'abuse', 'forced', 'rape', 'shooting',
           'violence', 'harassing', 'threat', 'threatening', 'threats', 'resist', 'resisting',
           'destruction', 'weapons', 'gun', 'knife', 'armed', 'deadly', 'drunkenness',
           'bomb', 'bombing', 'influence', 'looting', 'disorderly', 'force', 'forcible',
           'fighting', 'injuries', 'nuisance', 'homicide', 'alcohol', 'rape', 'mayhem',
           'abuse', 'cruelty', 'lewd', 'molest', 'disturbing']
       

sf_data['COP'] = sf_data['descript'].apply(lambda x: any(word in x for word in cop_words) )
sf_data['COP'].value_counts()



### Create new crime variable 'violant' for incidents in the categories assault, rape, domestic violence

violent_cats = ['assault', 'sex offenses, forcible', 'secondary codes']
sf_data['violent'] = sf_data['category'].apply(lambda x: x in violent_cats)
sf_data['violent'].value_counts()

### Create new crime variable southern_assaults
Just the assault incidents from the Southern Police District

sf_data[sf_data['category']== 'assault'].groupby(['pddistrict'])[['pdid']].count().sort_values('pdid', ascending=False)

#sf_data['southern_assaults'] = sf_data[['category','pddistrict']].apply(lambda x, y: (x=='assault') & (y=='SOUTHERN'), axis=1)

#df[['one','two']].apply(sum, axis=1)

#sf_data['southern_assaults'] = np.any((sf_data['category']=='assault') , (sf_data['pddistrict']=='SOUTHERN'))
#sf_data['southern_assaults'].value_counts()

### Create new crime variable 'gun' for incidents where a Gun was used
- Just wanted to know
- May be useful for other analysis

sf_data['gun'] = sf_data['descript'].apply(lambda x: x.find('gun') != -1 )
sf_data['gun'].value_counts()


sf_data.head(2)

## Consolidate into daily records
[[back to top](#Sections)]

Group the incidents by day and count the Number of incidents and the sum of the crime_level

### Group by the day
- Aggregate by the sum and count for my measuring variables
- To group by day and shift, uncomment code marked #SHIFT

day_group = sf_data.groupby(['date'])[['crime_level','weather_crime', 'violent', 'COP', 'gun']].agg(['sum', 'count'])


#SHIFT day_group = sf_data.groupby(['date','shift'])[['crime_level']].agg(['sum', 'count'])
day_group.head()

# unstack to bring shift from rows to columns
#SHIFT day_group = day_group.unstack(level=-1)
#SHIFT day_group.head()

day_group.columns.values

# Reduce column levels
day_group.columns = ['_'.join(col).strip() for col in day_group.columns.values]
day_group.head(2)

# drop the weather_crime_count and other count fields - they are just the count of all records
day_group.drop(['weather_crime_count', 'COP_count', 'gun_count', 'violent_count'], axis=1, inplace=True)

# Rename columns to avoid confusion
cols = ['crime_level_sum', 'crime_count', 'weather_crime_count', 'violent_count', 'COP_count', 'gun_crime_count']
day_group.columns = cols
day_group.head(2)

# Sum up the 3 shift info into day totals
#SHIFT day_group['crime_level_sum_day'] = day_group['crime_level_sum_shift_1'] + 
#                                   day_group['crime_level_sum_shift_2'] + 
#                                   day_group['crime_level_sum_shift_3']
#day_group['crime_level_count_day'] = day_group['crime_level_count_shift_1'] + 
#                                     day_group['crime_level_count_shift_2'] + 
#                                     day_group['crime_level_count_shift_3']        
#day_group.head(2)

## Add in the other fields that are not crime rate
Features that are needed for further analysis
- day, month, year and dayofweek

day_group_static = sf_data.groupby(['date'])[['dayofweek','day', 'month', 'year']].min()
day_group_static.head()

### Merge the two data Frames 

data = pd.concat([day_group, day_group_static], axis=1, join_axes=[day_group.index])
data.head()

### Review Data

data.describe()

#### minimum of 2 crimes looks bad

data.sort_values('crime_count', ascending=True).head(10)

#### Remove the two days that appear to be missing data
- there were 2 days that only had 2 and 8 incidents. 
- remove them since there must be missing data from those 2 days.

data = data[data['crime_count'] > 100]
data.shape

data.describe()

## Write final data to file
[[back to top](#Sections)]

data.to_csv('sf_crime_clean.csv')
