==> 3_clean_weather_combine_with_crime.ipynb <==
# Clean Historical Weather data and combine with Crime data

- Jim Haskin

- GA-Data Science
- Dec 2015

- 2/17/2016

## Method
- I collected historical weather data for the San Franciso area from January 2003 until December 2015. The data comes from Weather Underground. http://www.wunderground.com/history/
- I cleaned the data.
- To that data I added new features, such as the phase of the moon, that I wanted to investigate.
- I then combined the weather data with the crime data to form the final data file that is used for the analysis and modeling.


## Sections

- [Data Source](#Data-source)
- [Clean features](#Clean-features)
- [New features](#New-features)
- [Combine Weather and Crime data](#Combine-Weather-and-Crime-data)
- [qq](#qq)

import pandas as pd
import numpy as np
import seaborn as sb
%matplotlib inline
import ephem as ep
from datetime import timedelta
from sf_sun_moon import sf_sun_moon
from pywws import conversions as cv

## Data source

[[back to top](#Sections)]

- Data downloaded from WeatherUnderground by year and consolidated into a single file.
- sf_weather_2003_2015.csv
- http://www.wunderground.com/history/

FieldName|Type|Description                             
---------------|------------|---------------------
pst|string|Date in format : 2003-1-1
max_temperaturef|int|High Temperature for the day in degrees F
mean_temperaturef|int|Mean Temperature for the day in degrees F
min_temperaturef|int|Low Temperature for the day in degrees F
max_dew_pointf|int|High Dew Point for the day in degrees F 
meandew_pointf|int|Mean Dew Point for the day in degrees F
min_dewpointf|int|Low Dew Point for the day in degrees F
max_humidity|int|Maximum Humidity for the day in percentage 
mean_humidity|int|Mean Humidity for the day in percentage
min_humidity|int|Minimum Humidity for the day in percentage
max_sea_level_pressurein|float|High Sea Level for the day in inches
mean_sea_level_pressurein|float|Mean Sea Level for the day in inches
min_sea_level_pressurein|float|Low Sea Level for the day in inches
max_visibilitymiles|int|Maximum Visibility in miles
mean_visibilitymiles|int|Mean Visibility in miles
min_visibilitymiles|int|Minimum Visibility in miles
max_wind_speedmph|int|Maximum maintained Wind Speed in mph
mean_wind_speedmph|int|Mean maintained Wind Speed in mph
max_gust_speedmph|double|Maximum Wind Gust in mph
precipitationin|string|Amount is precipitation in inches
cloudcover|int|Cloud Cover, numeric ranges from 0 to 8
events|string|Significant Events such as 'Rain', 'Fog'
winddirdegrees<br_/>|string|General wind direction, numeric 0 359

!head -n 3 sf_weather_2003_2015.csv

### Read in data

w_data = pd.read_csv('sf_weather_2003_2015.csv')
w_data.head(2)

### Simplify feature labels

labels = ['pst', 'temp_max', 'temp_mean', 'temp_min', 'dew_point_max', 'dew_point_mean', 'dew_point_min',
          'humidity_max', 'humidity_mean', 'humidity_min', 'sl_pressure_max', 'sl_pressure_mean', 'sl_pressure_min',
          'visibility_max', 'visibility_mean', 'visibility_min', 'wind_speed_max', 'wind_speed_mean', 
          'wind_max_gust', 'precipitation', 'cloud_cover', 'events', 'wind_direction']

w_data.columns = labels
w_data.head(2)

### Investigate data

w_data.describe()

w_data.info()

### Observations
- 4747 records
- Date in string form
- most measurements in int or float
- precipitation, wind_direction not numeric (fix)
- missing 465 wind_max_gust ~10%
- missing 3581 events ~75%. But will assume that a null is interpreted as no event happened

## Clean features
[[back to top](#Sections)]

### Convert date to datetime

w_data['date'] = pd.to_datetime(w_data['pst'])

### Clean up wind_direction number
- line break was appended to number

w_data['wind_direction'] = w_data['wind_direction'].str.replace('<br />','').astype('int')
#w_data['wind_direction'] = w_data['wind_direction'].replace('<br_/>','',regex=True).astype('int')
#w_data['wind_direction'].astype(int, copy=True)

### Clean precipitation. 
- Has 'T' for trace instead of number
- 'T' forces the field to string type
- Replace trace with small numeric (0.005, half of lowest recorded)

w_data['precipitation'].value_counts().head(5)

w_data['precipitation'].dtype

w_data['precipitation'] = w_data['precipitation'].str.replace('T','.005').astype('float')

### Cloud cover
- in historical weather underground cloud cover ranges from 0 to 8.( 4 mean 4/8) The cloud cover number I will use later for the forecast is in %
- convert to %

w_data['cloud_cover'] = (w_data['cloud_cover']*100/8).astype('int')

## Catagoricals
### events
- records rain, fog, thunderstorm events
- will assume that nulls are days 'No' event occured
- since low occuring events are combinations of other events, consider manipulation to eliminate sparce events

w_data['events'].fillna('No', inplace=True)
#w_data['events'] = w_data['events'].astype('category')

w_data['events'].value_counts()

### Review data

w_data.info()

## New features
[[back to top](#Sections)]
- These are new features that may help in the prediction of the crime level.
- Some are calculated based on the day. (Length of sun, etc.)
- Some features interact with other, such as wind chill

### Sun and Moon information
- call  sf_sun_moon.py subroutine to get:
- length of minutes of Sunlight for each day
- length of minutes of Moonlight for each day
- Phase of Moon

PROBLEM - values for moon length are over 24 hours check out - do not use

w_data['sun_length'], w_data['moon_phase'] = zip(*w_data['pst'].apply(sf_sun_moon))
#w_data['sun_length'], w_data['moon_length'], w_data['moon_phase'] = zip(*w_data['pst'].apply(sf_sun_moon))




### Daily Sun Level
- Use the cloud cover number and length of sun light to calculate a new feature, 'sun_units'

w_data['sun_units'] = w_data['sun_length'] * (1.0 - (w_data['cloud_cover']  * .01))

### Dew Point
- use pywws.conversions module to get following features

#pywws.conversions.dew_point(temp, hum)[source]
w_data.loc[:,'dew_point_mean'] = w_data.apply( lambda x : cv.dew_point(x['temp_mean'], x['humidity_mean']), axis=1)

### Wind chill and Real feel
- Wind and humidity levels can effect how warm or cold you actually feel and effect peoples behavior .
- Create feature for wind chill (temp_mean/wind_speed_mean)
- Create feature for humidity/temp (temp_mean?????humidity_mean)

#pywws.conversions.wind_chill(temp, wind)
w_data.loc[:,'wind_chill_mean'] = w_data.apply( lambda x : cv.wind_chill(x['temp_mean'], x['wind_speed_mean']), axis=1)
w_data.loc[:,'wind_chill_max'] = w_data.apply( lambda x : cv.wind_chill(x['temp_max'], x['wind_speed_max']), axis=1)

#pywws.conversions.usaheatindex(temp, humidity, dew)
w_data.loc[:,'heat_index'] = w_data.apply( lambda x : cv.usaheatindex(x['temp_mean'], x['humidity_mean'], x['dew_point_mean']), axis=1)
#pywws.conversions.apparent_temp(temp, rh, wind)
w_data.loc[:,'temp_apparent'] = w_data.apply( lambda x : cv.apparent_temp(x['temp_mean'], x['humidity_mean'], x['wind_speed_mean']), axis=1)

### Temperature and Humidy swings during the day
- Drastic changes in temp, pressure and humidity may effect peoples behavior.
- Create temp_delta
- Create humidity_delta 

w_data.insert(4, 'temp_delta', w_data['temp_max'] - w_data['temp_min'])

### Possible features still to add 
- daily deltas(swings) - humidity, pressure
- day to day deltas of average/max/min - temp, humidity, pressure


w_data.head()

## Index, Remove unwanted fields and save final data to csv file

# Set index to the data
w_data.set_index(w_data['date'], inplace=True)
w_data.head(2)

data = w_data.drop(['pst', 'visibility_max', 'visibility_mean', 'visibility_min'], axis=1)
data.head()

# Save only weather data to file for backup
data.to_csv('sf_weather_clean.csv')

## Combine Weather and Crime data 
[[back to top](#Sections)]

# Read in crime data and rename the weather data
crime = pd.read_csv('sf_crime_clean.csv', index_col=0)
weather = data
#weather = pd.read_csv('sf_weather_clean.csv', index_col=0)

# Merge data and remove duplicate data field
merge_data = crime.merge(weather, left_index=True, right_index=True)
data = merge_data.drop(['date'], axis=1)
merge_data.head(10)

# write final data to file
data.to_csv('sf_crime_weather.csv')
